Extract UI-Header into new file Layout.tsx
